+++ CHANGELOG.md
@@ -7,5 +7,7 @@
 - Adicionando atributo "width" na classe Plataforma
 - Adicionando metodo "pushEntidades" na classe Fase
 - Adicionando metodos "createMorcego", "createGhoul", "createPlataforma" e "createEspinho" na classe Fase
-- Adicionado metodo "deal" na classe Entidade
-- Adicionado metodo "receive" na classe Personagem
+
+- Adicionado metodo "deal" & "receive" na classe Entidade, Inimigo, Jogador
+- Adicionado metodo "deal" na classe Obstaculo, Espinho
+- Alteracoes nos metodos na classe CollisionManager